src/deploy/requirements.txt
azure-functions
azure-eventhub